+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Innmind\BlackBox\PHPUnit\Framework\TestCase`
 - `Innmind\BlackBox\PHPUnit\Load`
 - `Innmind\BlackBox\Application::disableMemoryLimit()`
+- `Innmind\BlackBox\Runner\Assert::matches()`
 
 ### Fixed